Website Updates
I updated the head to have a header, aside and article. The header of which was given a navigation tab with links to both the Home (or index) and the Services pages. I also added in reviews from customers under a comment section that was put in blockquotes. Lastly, I added a comment that tried to be reasonably detailed before each question/task that I put into the document.