Finished the webpages.
I started first by cleaning up my navigation and then adding in the reviews with blockquotes and cites to boot. Then I added the main heading for the website which includes a linked email address and phone #. Lastly, I went into jtc_services_txt.html and added all the headings and paragraphs along with an image and the footer that included the location's address.